[refactor]: SavedScreen content들 component로 분리(#26)
