+++ app/src/main/java/com/texthip/thip/ui/myPage/SavedScreen.kt
@@ -8,8 +8,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Tab
@@ -17,7 +15,6 @@ import androidx.compose.material3.TabRow
 import androidx.compose.material3.TabRowDefaults.tabIndicatorOffset
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
@@ -26,20 +23,15 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.texthip.thip.R
-import com.texthip.thip.ui.common.cards.CardBookList
 import com.texthip.thip.ui.common.topappbar.DefaultTopAppBar
-import com.texthip.thip.ui.myPage.component.SavedFeedCard
-import com.texthip.thip.ui.myPage.viewmodel.SavedBookViewModel
-import com.texthip.thip.ui.myPage.viewmodel.SavedFeedViewModel
+import com.texthip.thip.ui.myPage.component.BookContent
+import com.texthip.thip.ui.myPage.component.FeedContent
 import com.texthip.thip.ui.theme.Black
-import com.texthip.thip.ui.theme.Grey02
 import com.texthip.thip.ui.theme.ThipTheme.colors
 import com.texthip.thip.ui.theme.ThipTheme.typography
 import com.texthip.thip.ui.theme.White
@@ -112,43 +104,6 @@ fun SavedScreen() {
     }
 }
 
-@Composable
-fun FeedContent(viewModel: SavedFeedViewModel = viewModel()) {
-    val feedList by viewModel.feeds.collectAsState()
-
-    LazyColumn {
-        items(feedList, key = { it.id }) { feed ->
-            val bookImagePainter = feed.imageUrl?.let { painterResource(it) }
-            val profileImagePainter = feed.userProfileImage?.let { painterResource(it) }
-
-            SavedFeedCard(
-                feedItem = feed,
-                bookImage = bookImagePainter,
-                profileImage = profileImagePainter,
-                onBookmarkClick = { viewModel.toggleBookmark(feed.id) },
-                onLikeClick = { viewModel.toggleLike(feed.id) }
-            )
-        }
-    }
-}
-
-@Composable
-fun BookContent(viewModel: SavedBookViewModel = viewModel()) {
-    val books by viewModel.bookList.collectAsState()
-
-    LazyColumn {
-        items(items = books, key = { it.id }) { book ->
-            CardBookList(
-                title = book.title,
-                author = book.author,
-                imageRes = null,
-                publisher = book.publisher,
-                isBookmarked = book.isSaved,
-                onBookmarkClick = { viewModel.toggleBookmark(book.id) }
-            )
-        }
-    }
-}
 
 @Preview
 @Composable
