+++ app/src/main/java/com/texthip/thip/ui/myPage/component/FeedContent.kt
@@ -1,0 +1,30 @@
+package com.texthip.thip.ui.myPage.component
+
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.res.painterResource
+import androidx.lifecycle.viewmodel.compose.viewModel
+import com.texthip.thip.ui.myPage.viewmodel.SavedFeedViewModel
+
+@Composable
+fun FeedContent(viewModel: SavedFeedViewModel = viewModel()) {
+    val feedList by viewModel.feeds.collectAsState()
+
+    LazyColumn {
+        items(feedList, key = { it.id }) { feed ->
+            val bookImagePainter = feed.imageUrl?.let { painterResource(it) }
+            val profileImagePainter = feed.userProfileImage?.let { painterResource(it) }
+
+            SavedFeedCard(
+                feedItem = feed,
+                bookImage = bookImagePainter,
+                profileImage = profileImagePainter,
+                onBookmarkClick = { viewModel.toggleBookmark(feed.id) },
+                onLikeClick = { viewModel.toggleLike(feed.id) }
+            )
+        }
+    }
+}
